+++ kadai1/shuheiktgw/converter-cli/cli.go
@@ -1,11 +1,11 @@
-package main
+package converter_cli
 
 import (
 	"flag"
 	"fmt"
 	"io"
 
-	"github.com/shuheiktgw/dojo4/kadai1/converter-cli/converter"
+	"github.com/shuheiktgw/dojo4/kadai1/shuheiktgw/converter-cli/converter"
 )
 
 const (
