+++ kadai1/shuheiktgw/converter-cli/cli_test.go
@@ -1,4 +1,4 @@
-package main
+package converter_cli
 
 import (
 	"bytes"
