+++ kadai1/shuheiktgw/converter-cli/converter/testdata/gif-image.gif
