+++ kadai1/shuheiktgw/converter-cli/converter/testdata/jpeg-image1.jpeg
