+++ kadai1/shuheiktgw/converter-cli/converter/testdata/jpg-image1.jpg
