+++ kadai1/shuheiktgw/converter-cli/converter/testdata/png-image.png
