+++ kadai1/shuheiktgw/converter-cli/converter/testdata/png-image1.png
