+++ kadai1/shuheiktgw/converter-cli/main.go
@@ -6,7 +6,7 @@
 //   --from value, -f value  specifies a image extension converted from (default: .jpg)
 //   --to value, -t value    specifies a image extension converted to (default: .png)
 //   --help, -h              prints out help
-package main
+package converter_cli
 
 import (
 	"os"
