+++ kadai1/shuheiktgw/converter-cli/testdata/jpeg-image.jpeg
